+++ CHANGELOG.md
@@ -4,12 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
 ### Added
 
 ### Changed
 
 ### Fixed
+
+## [1.0.3] - 2022-07-03
+### Added
+- Better logs for GPU detection.
+
+### Changed
+- Move output.log to a fixed relative location. 
+
+### Fixed
+- Fix for AMD hardware encoding.
 
 ## [1.0.2] - 2022-06-28
 ### Changed